+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [49.4.0] (2026-08-13)
 
 ### Enhanced Experience Bar
 
@@ -2748,7 +2748,8 @@ Refreshes the material list and recipe tree when you've crafted or gathered an i
 
 Opens duty finder for the duty you clicked on in the Wondrous Tails Journal.
 
-[unreleased]: https://github.com/Haselnussbomber/HaselTweaks/compare/v49.3.0...main
+[unreleased]: https://github.com/Haselnussbomber/HaselTweaks/compare/v49.4.0...main
+[49.4.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v49.3.0...v49.4.0
 [49.3.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v49.2.4...v49.3.0
 [49.2.4]: https://github.com/Haselnussbomber/HaselTweaks/compare/v49.2.3...v49.2.4
 [49.2.3]: https://github.com/Haselnussbomber/HaselTweaks/compare/v49.2.2...v49.2.3